fix sniffer
