+++ lib/site-inspector.rb
@@ -13,7 +13,7 @@ require_relative 'site-inspector/endpoint'
 require_relative 'site-inspector/endpoint/dns'
 require_relative 'site-inspector/endpoint/headers'
 require_relative 'site-inspector/endpoint/hsts'
-require_relative 'site-inspector/sniffer'
+require_relative 'site-inspector/endpoint/sniffer'
 require_relative 'site-inspector/version'
 
 class SiteInspector
